Modularize loadV2V3Contract
