+++ src/utils/v2v3/contractLoaders/PublicResolver.ts
@@ -1,0 +1,11 @@
+import { goerliPublicResolver } from 'constants/contracts/goerli/PublicResolver'
+import { mainnetPublicResolver } from 'constants/contracts/mainnet/PublicResolver'
+import { rinkebyPublicResolver } from 'constants/contracts/rinkeby/PublicResolver'
+import { NetworkName } from 'models/network-name'
+
+export const loadPublicResolverContract = (network: NetworkName) => {
+  // ENS contracts package currently doesn't include rinkeby information, and ABI contains errors
+  if (network === NetworkName.mainnet) return mainnetPublicResolver
+  if (network === NetworkName.rinkeby) return rinkebyPublicResolver
+  if (network === NetworkName.goerli) return goerliPublicResolver
+}
